requirements.txt
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.2.0
torch==2.0.0
tqdm==4.64.1
